Prod Release 24/07/2024

+++ runner/src/stream-handler/stream-handler.ts
@@ -2,16 +2,19 @@ import path from 'path';
 import { Worker, isMainThread } from 'worker_threads';
 
 import { registerWorkerMetrics, deregisterWorkerMetrics } from '../metrics';
-import Indexer from '../indexer';
 import LogEntry from '../indexer-meta/log-entry';
 import logger from '../logger';
 
 import type IndexerConfig from '../indexer-config';
+import IndexerMeta, { IndexerStatus } from '../indexer-meta';
+import assert from 'assert';
+import Provisioner from '../provisioner';
+import { type PostgresConnectionParams } from '../pg-client';
 
 export enum WorkerMessageType {
-  METRICS = 'METRICS',
-  BLOCK_HEIGHT = 'BLOCK_HEIGHT',
-  EXECUTION_STATE = 'STATUS',
+  METRICS,
+  BLOCK_HEIGHT,
+  EXECUTION_STATE,
 }
 
 export interface WorkerMessage {
@@ -20,11 +23,13 @@ export interface WorkerMessage {
 }
 
 export enum ExecutionState {
+  UNSPECIFIED = 'UNSPECIFIED',
   RUNNING = 'RUNNING',
   FAILING = 'FAILING',
   WAITING = 'WAITING',
   STOPPED = 'STOPPED',
   STALLED = 'STALLED',
+  IDLE = 'IDLE',
 }
 
 interface ExecutorContext {
@@ -34,59 +39,79 @@ interface ExecutorContext {
 
 export default class StreamHandler {
   private readonly logger: typeof logger;
-  private readonly worker: Worker;
+  private worker: Worker | undefined;
   public readonly executorContext: ExecutorContext;
+  private indexerMeta: IndexerMeta | undefined;
 
   constructor (
     public readonly indexerConfig: IndexerConfig,
   ) {
+    this.logger = logger.child({ accountId: indexerConfig.accountId, functionName: indexerConfig.functionName, service: this.constructor.name });
+
+    this.executorContext = {
+      executionState: ExecutionState.IDLE,
+      block_height: indexerConfig.version,
+    };
+  }
+
+  async start (): Promise<void> {
     if (isMainThread) {
-      this.logger = logger.child({ accountId: indexerConfig.accountId, functionName: indexerConfig.functionName, service: this.constructor.name });
+      try {
+        const provisioner = new Provisioner();
+        const databaseConnectionParams: PostgresConnectionParams = await provisioner.getPgBouncerConnectionParameters(this.indexerConfig.hasuraRoleName());
 
-      this.worker = new Worker(path.join(__dirname, 'worker.js'), {
-        workerData: {
-          indexerConfigData: indexerConfig.toObject(),
-        },
-      });
-      this.executorContext = {
-        executionState: ExecutionState.RUNNING,
-        block_height: indexerConfig.version,
-      };
+        this.indexerMeta = new IndexerMeta(this.indexerConfig, databaseConnectionParams);
+        this.worker = new Worker(path.join(__dirname, 'worker.js'), {
+          workerData: {
+            indexerConfigData: this.indexerConfig.toObject(),
+            databaseConnectionParams,
+          },
+        });
 
-      this.worker.on('message', this.handleMessage.bind(this));
-      this.worker.on('error', this.handleError.bind(this));
+        this.worker.on('message', this.handleMessage.bind(this));
+        this.worker.on('error', this.handleError.bind(this));
+
+        this.executorContext.executionState = ExecutionState.RUNNING;
+      } catch (error: any) {
+        const errorContent = error instanceof Error ? error.toString() : JSON.stringify(error);
+        this.logger.error('Terminating thread', error);
+        this.executorContext.executionState = ExecutionState.STALLED;
+        throw new Error(`Failed to start Indexer: ${errorContent}`);
+      }
     } else {
-      throw new Error('StreamHandler should not be instantiated in a worker thread');
+      throw new Error('StreamHandler should not be started in a worker thread');
     }
   }
 
   async stop (): Promise<void> {
-    deregisterWorkerMetrics(this.worker.threadId);
-
+    if (this.worker) {
+      deregisterWorkerMetrics(this.worker.threadId);
+      await this.worker.terminate();
+    }
     this.executorContext.executionState = ExecutionState.STOPPED;
-
-    await this.worker.terminate();
   }
 
   private handleError (error: Error): void {
     this.logger.error('Terminating thread', error);
     this.executorContext.executionState = ExecutionState.STALLED;
 
-    const indexer = new Indexer(this.indexerConfig);
-    indexer.setStoppedStatus().catch((e) => {
-      this.logger.error('Failed to set stopped status for indexer', e);
-    });
-    const errorContent = error instanceof Error ? error.toString() : JSON.stringify(error);
-    const streamErrorLogEntry = LogEntry.systemError(`Encountered error processing stream: ${this.indexerConfig.redisStreamKey}, terminating thread\n${errorContent}`, this.executorContext.block_height);
-
-    indexer.writeCrashedWorkerLog(streamErrorLogEntry)
-      .catch((e) => {
-        this.logger.error('Failed to write failure log for stream', e);
+    if (this.indexerMeta) {
+      this.indexerMeta.setStatus(IndexerStatus.STOPPED).catch((e: Error) => {
+        this.logger.error('Failed to set stopped status for indexer', e);
       });
+      const errorContent = error instanceof Error ? error.toString() : JSON.stringify(error);
+      const streamErrorLogEntry = LogEntry.systemError(`Encountered error processing stream: ${this.indexerConfig.redisStreamKey}, terminating thread\n${errorContent}`, this.executorContext.block_height);
+      this.indexerMeta.writeLogs([streamErrorLogEntry])
+        .catch((e) => {
+          this.logger.error('Failed to write failure log for stream', e);
+        });
+    }
 
-    this.worker.terminate().catch(() => {
-      this.logger.error('Failed to terminate thread for stream');
-    });
+    if (this.worker) {
+      this.worker.terminate().catch(() => {
+        this.logger.error('Failed to terminate thread for stream');
+      });
+    }
   }
 
   private handleMessage (message: WorkerMessage): void {
@@ -98,6 +123,7 @@ export default class StreamHandler {
         this.executorContext.block_height = message.data;
         break;
       case WorkerMessageType.METRICS:
+        assert(this.worker, 'Worker is not initialized');
         registerWorkerMetrics(this.worker.threadId, message.data);
         break;
     }
